+++ src/main/java/chatsy/exceptions/ChatsyException.java
@@ -1,4 +1,4 @@
-package main.java.chatbot.exceptions;
+package chatsy.exceptions;
 
 public class ChatsyException extends Exception {
     @Override
