+++ src/main/java/chatsy/exceptions/FileCorruptedException.java
@@ -1,4 +1,4 @@
-package main.java.chatbot.exceptions;
+package chatsy.exceptions;
 
 public class FileCorruptedException extends LocalFileException {
     private String filePath;
