+++ src/main/java/chatsy/exceptions/InvalidTaskIndexException.java
@@ -1,4 +1,4 @@
-package main.java.chatbot.exceptions;
+package chatsy.exceptions;
 
 public class InvalidTaskIndexException extends ChatsyException {
     @Override
