+++ src/main/java/chatsy/exceptions/LocalFileException.java
@@ -1,4 +1,4 @@
-package main.java.chatbot.exceptions;
+package chatsy.exceptions;
 
 public class LocalFileException extends ChatsyException {
     private String filePath;
