+++ src/main/java/chatsy/tasks/DeadlineTask.java
@@ -1,4 +1,4 @@
-package main.java.chatbot.tasks;
+package chatsy.tasks;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
